third_party/WebKit/LayoutTests/fast/dom/shadow/shadowroot-clonenode-expected.txt

CONSOLE WARNING: line 11: Element.createShadowRoot is deprecated and will be removed in M73, around March 2019. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 for more details.
Calling ShadowRoot.cloneNode() should throw a NotSupportedError exception.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS shadowRoot.cloneNode() threw exception NotSupportedError: Failed to execute 'cloneNode' on 'Node': ShadowRoot nodes are not clonable..
PASS successfullyParsed is true

TEST COMPLETE
